Inspector v2: Fix pane resizing (#17305)
Two issues were reported on the forum related to pane resizing:
1. There is some bug where sometimes resizing doesn't work (you get a 🚫
cursor), and then it gets in a bad state that takes a few resize retries
to recover from.
2. The Fluent package with useResizeHandle has a peer dependency version
of React that excludes React 19, so the Inspector v2 package fails to
install in projects already using React 19.

I've reported both of these Fluent issues, but in the meantime I just
created a custom version of useResizeHandle that is far less feature
rich and generally not as good, but should be fine for the time being.
Easy to swap back over to the Fluent one when the issues are fixed.