SPLATFileLoader: add user-provided fflate instance to options (#17310)
In SPLATFileLoader, a fflate instance is loaded dynamically when being
used, and the loaded lib is exposed to global context by default. To be
able to avoid this, a `fflate` option has been added to
`SPLATLoadingOptions`, and when assigned, loader would use this instead
of loading a fflate to global.

Forum post:
<https://forum.babylonjs.com/t/splatfileloader-possible-to-avoid-global-vars-and-externals/60927>