Inspector v2: Add Dispose button for any disposable in the properties pane (#17317)
This is in response to
https://forum.babylonjs.com/t/introducing-inspector-v2/60937/54

Instead of selectively adding a Dispose button in the properties pane
for some entities, we now add a Dispose button to the bottom of the
General section for any entity that implements `IDisposable` (e.g. has a
`dispose` function).

Also a tiny bit of cleanup from a recent PR that uses `Collapse` instead
of conditionally rendering a component, and uses Title Case for property
labels.