FrameGraph: Implement TAA as a post process (#17326)
This is a better implementation than the previous one (which was
implemented as a renderer) because it is now possible to insert other
processes (such as post-processes) between scene rendering and the TAA
effect (which is also possible with the standard TAA implementation,
outside of frame graphs).

I also added support for the new `clampHistory` and `reprojectHistory`
properties in the new TAA frame graph task.

There is one breaking change in this PR:
`Material.setVertexOutputInvariant` has been renamed to
`Material.isVertexOutputInvariant`. I don't think this is a problem, as
this property is fairly new and very specific, and I don't think anyone
is using it yet.