Smart Filter Editor: fix scrollbar to prevent frequent canvas resizes (#17335)
The preview was included in the scroll region, unlike the other editors,
so when it got a vertical scrollbar, that changed the width of the
preview which caused the engine to resize. Now SFE behaves like the
other editors, and the preview doesn't scroll off screen and isn't
affected by the scrollbar.

Co-authored-by: AmoebaChant <kevbrown@microsoft.com>