Playground: Fix tab handling regression from Monaco update (#17337)
The Playground [Editor
v2](https://github.com/BabylonJS/Babylon.js/pull/17216) updated Monaco
which regressed the [workaround
](https://github.com/BabylonJS/Babylon.js/pull/16741/)for tab handling,
where when Inspector v2 is open, pressing the tab key while focus is in
the code editor moves focus instead of inserting a tab character. To
work around this, we added the contentEditable attribute to the actual
Monaco html element that has keyboard focus when editing text, but that
element has changed in the newer Monaco, so the workaround needs to be
updated. Also, the Editor v2 PR removed the comment explaining the
workaround, so I just re-added it from the original change.