useLargeWorldRendering on engine enables floatingOrigin and highMatrixPrecision. Add ability to useFloatingOrigin per-scene (#17334)
This PR solves 2 major bugs in floatingoriginmode. 
- Previously I was updating the creationOptions object but that didn't
actually update the underlying performanceConfigurator's precision, so
we were getting CPU-side calculation imprecision.
- Additionally, the floatingOriginOffset value was tied to the most
recently created scene's active camera, and didn't support multi-scene.

After this change:
- engine creation **useLargeWorldRendering** will both set
useHighPrecisionMatrix on engine and set all scenes to useFloatingOrigin
upon creation
- scene creation **useFloatingOrigin** can be used instead of the above
if user wants only some scenes to useFloatingOrigin (doing this can
improve perf for scenes that are not large worlds as we skip matrix
offsetting). If this option is used, must be used in conjunction with
engine's useHighPrecisionMatrix to achieve full precision)
- offset will always be tied to the current rendering scene's active
camera

Tested with this playground [#P3E9YP#255
](https://playground.babylonjs.com/?snapshot=refs/pull/17334/merge#P3E9YP#255)
which enables swapping between scenes and testing combinations of
floatingOrigin on / off

---------

Co-authored-by: Georgina Halpern <gehalper@microsoft.com>