Single-pass IBL shadow voxelization for WebGPU (#17270)
This changes the voxelization strategy for WebGPU to allow it to be done
in a single pass (though, generation of mips still takes additional
steps). This is possible using `textureStore` to write to arbitrary
texels in a 3D texture. It also uses vertex pulling to correctly
determine the provoking vertex for each triangle and retrieve adjacent
vertices so that we can calculate the triangle's normal. This is used to
swizzle the xyz components to maximize the rasterized area and not miss
voxels.

Anywho, this makes updating shadows for animated meshes much more
plausible. I'm not sure the best way to do this but currently, in my
playground, I'm doing this:
```
const frameSkip = 1;
let frame = 0;
scene.onAfterAnimationsObservable.add(() => {
    if (enableIblShadows) {
        if (frame == 0 && scene.animatables.find((anim) => !anim.paused)) {
            iblShadowsPipeline.updateVoxelization();
            iblShadowsPipeline.resetAccumulation();
        }
        if (frame === frameSkip) {
            frame = 0;
        } else {
            frame++;
        }
    }
});
```
Is there a better way to query whether there is currently an animation
happening?

---------

Co-authored-by: Mike Bond <mdcbond@gmail.com>