Smart Filters: fix over-aggressive optimizer validation error when local variable contains a uniform name (#17350)
If a local variable in a helper function contained a string matching the
decorated version of a uniform name, it would incorrectly think that the
helper function was using the uniform and return an error. For example,
if the uniform was "foo" the decorated name would be "_foo_". If the
function had a parameter called "foo_bar" it would become "_foo_bar_"
and then would incorrectly be confused for the uniform "_foo_" and the
error would be displayed. This fixes test to ensure there are non-word
character surrounding the match.

Co-authored-by: AmoebaChant <kevbrown@microsoft.com>