Merging textures on export (#17339)
For OpenPBR, all textures that can be combined in glTF can also be
assigned separately. Therefore, we need a way to easily merge textures
on export.

I've written a texture-merging utility and started implmenting it for
the anisotropy extensions as well as baseColor-alpha and metal-rough
texture output.

I'd like feedback about this approach before I continue further. Please
and thank you.