glTFLoaderAnimation: do not clone animation keys (#17352)
This commit skips animation keys cloning by setting `dontClone=true`
when calling `Animation.setKeys`. This should reduce peak heap
allocation and GC pressure when loading animation-heavy models.
In GLTF loader, keys are created in callback in
`_loadAnimationChannelFromTargetInfoAsync`, and passed to
`propertyInfo.buildAnimations`, and never escaped from the scope, so it
should be safe to do this change.

See also the forum post
<https://forum.babylonjs.com/t/gltf-loader-do-not-clone-animation-keys/60903>