[FloatingOrigin] Web GPU Fix (#17356)
in FloatingOriginMode, we override effect.setMatrix to apply an offset.
This in turn calls pipelineContext.setMatrix, which in
webGPUPipelineContext will call ubo.updateMatrix

previously, this would apply offset again and cause incorrect
positioning. The fix is to avoid double offset by existing early if the
passed in matrix is === the most recently offset Temp matrix. To avoid
cases where TempFinalMat happens to be equal values, I also move the
updateFlag set to after the early out check.

Co-authored-by: Georgina Halpern <gehalper@microsoft.com>