Fix intermittent build issue with build:assets:smart-filters (#17351)
The build:assets:smart-filters script was using tsx to run the
buildShader.ts script, but in some versions of node, it was unable to
resolve dependencies on other packages.

For example:
SyntaxError: The requested module 'core/Materials/effectRenderer.js'
does not provide an export named 'EffectWrapper'

With this change, that script now uses the public buildShader.js
instead, eliminating the dependency resolution issue. The only downside
is that working on the Smart Filter code will require building the
public package first - and the same with unit tests. This is about a 20
second additional cost, and doesn't affect the other dev inner loops.

Co-authored-by: AmoebaChant <kevbrown@microsoft.com>