Inspector v2: umd package (#17357)
This is the initial implementation of the Inspector v2 umd package. We
need this for a few reasons:
1. There was one ask on the forum already for being able to use
Inspector directly in a web page without bundling the code. The only
solution we have for this right now is umd as we don't have bundled esm
packages, and using packages with thousands of loose files directly in a
page via esm is impractical due to the huge import map configuration
that would be needed.
2. Right now the versioning strategy used in Playground and Sandbox are
based on umd bundles.
3. We need umd for back compat.

The inspector v2 js bundle has a -v2 in the name, so it will not
conflict with Inspector v2 when deployed to the Babylon CDN.

After this change goes in, I will have to update the pipeline to have
the extra steps for publishing Inspector v2 as
babylonjs-inspector@preview (basically the same as what has already been
done with @babylonjs/inspector@preview).