Inspector v2: umd cdn (#17379)
This is a follow up to my initial umd package PR:
https://github.com/BabylonJS/Babylon.js/pull/17357

In the initial PR, the path I went was to redirect `inspector` to
`inspector-v2` in the context of the Inspector v2 build, since they
ultimately are the same package with different versions (v2 having the
`-prefix` suffix). This was simpler, but it turns out doesn't work with
the way we publish umd bundles to our cdn. For that, we need unique
entries in the package mappings, so now I've added a separate entry for
inspector-v2. This gets used by the prepareSnapshot script to determine
what gets copied to the cdn. The way I set it up is the inspector v2
bundle gets copied to the inspector directory, but with a new unique
name (babylon.inspector-v2.bundle.js). I tested the snapshot copy
locally, but in the process I noticed that we were copying
webpack.config.js files to the cdn (because they match the *.js glob).
For example: https://cdn.babylonjs.com/webpack.config.js. @docEdub - we
should probably try to figure out how to remove all these
webpack.config.js files from the cdn.

For now I'm making this PR draft so that I can verify the snapshot cdn
functionality before merging, and also I don't want to touch any cdn
publishing related stuff on a Friday afternoon. :)