SSAO2 camera fovMode support + Initialization with Ortho Camera Fix (#17388)
This PR complements #17383

### SSAO camera fovMode support 

This is the second topic I posted on the forum. 

https://forum.babylonjs.com/t/ssao2-does-not-behave-correctly-with-dynamically-changing-camera-parameters/61245

### Initialization with Ortho Camera Fix
First, the code in #17383 does not function correctly during the initial
camera mode initialization.
```typescript
this._originalColorPostProcess.onBeforeRenderObservable.add(() => {
    const camera = this._scene.activeCamera;
    if (camera && this._currentCameraMode !== camera.mode) {
        this._currentCameraMode = camera.mode;
        this._thinSSAORenderingPipeline._ssaoPostProcess.updateEffect();
    }
});
```
Because at the initial point when this callback executes,
`this._thinSSAORenderingPipeline._ssaoPostProcess` is null.

Therefore, the camera's initial projection `mode` is ignored, and since
the ORTHOGRAPHIC_CAMERA define is never set, it only functions correctly
for perspective cameras.

To resolve this, initialize
`this._thinSSAORenderingPipeline._ssaoPostProcess.camera` with
`activeCamera` once at line 331 in `ssao2RenderingPipeline.ts`.