FrameGraph: add new task + several fixes (#17381)
In this PR:
* Added a new `FrameGraphComputeShaderTask` (+
`NodeRenderGraphComputeShaderBlock`) class to simplify the management of
compute shaders in frame graphs
* Added a new `customIsReady` property in `FrameGraphExecuteTask`
* Fixed directional and point shadows not working with frame graphs in
some cases
* Fixed `Mesh.ignoreCameraZ` not working with frame graphs
* Regarding this one, making `Mesh.ignoreCameraMaxZ` work in
`ObjectRenderer` may be difficult (it wasn't supported when the feature
has been implemented a few years ago) and bad for performance, because
it means the view/projection matrix changes during rendering, forcing us
to update the scene ubo, which is bad in WebGPU.
* Instead, I added a **ignoreCameraMaxZ** property in the `Camera`
class. I feel it's the right place to put it, and there's no performance
penalty to put it there. The catch is that all meshes rendered by the
camera will ignore **maxZ** when the property is set. That shouldn't be
a problem, though. What do you think? cc @sebavan, @deltakosh
* Fixed crashes in NRGE in some cases

~~Don't review it yet, it may not be complete!~~ Ready for review!