DumpData: Refactor for glTF export in BN (#17365)
This PR adds the glue code to enable DumpData usage in BN with the glTF
exporter.

Changes:
- Factor out core image encoding to a separate function matching
NativeEncoding's own exposed function name. This follows the current
nativeOverride pattern for module-level functions (see the MathHelper
class.)
- Remove the `bitmaprenderer` path. I initially planned for Native to
polyfill this, but that turned out to be not a good idea.
- Fix case in glTF exporter where exporting images to a specific format
could lead to invalid images if the browser or platform didn't support
it.