GPUPicker: Fix picking ID precision issues (#17404)
Resolves the issue:

https://forum.babylonjs.com/t/missing-doc-for-gpupicker-in-v8-31-3/61295/5

What I found while examining this issue is that

when outputting values received as uniform variables in the shader to
the screen,
the uniform values were displayed completely without error,

but values received as attributes exhibited errors.

I don't know the exact reason for this, as it is an internal WebGL
implementation detail.
Through experimentation, I found a safe `MaxPickingId` value usable
across WebGL, WebGL2, and WebGPU.
I confirmed that adding 0.5 at the point of casting from int resolves
this precision issue.