Support multi-object raycasting for havok (#17396)
We use multi-object raycasting to see if a trigger is in view, but it
can be useful in other ways too (for example, guns that shoot through
multiple enemies).

Let me know if this API is a good fit for Babylon.js. Another approach
could be to create a new `raycastMulti` function in `HavokPlugin`, and
reimplement `raycast` as a thin wrapper over it.

One possible concern is if a developer sets `maxQueryCollectorHits` to a
high number, it will affect *all* raycasts, not just multi-casts. We
could perhaps create a separate `queryCollector` for multi-cast use
only?