Inspector v2: debugLayer and back compat first pass (#17421)
This PR adds a simple `Scene.debugLayer` implementation, adds support
for additional Inspector v1 features and APIs, adds a little new
functionality to the core Inspector v2 layers to support some back
compat features, and clearly separates back compat functionality from
new functionality/API.

- Add a `layoutMode` to `ShellService` that allows inline or overlay
(side panes overlay the central content).
- Allow the `sidePaneRemapper` option for `ShellService` to return
`null`, meaning don't show the pane at all (used for Inspector v1
options that shows only Scene Explorer or only the rest of the right
side panes).
- Removed the `Inspector` class and all Inspector v1 related
functionality from the Inspector v2 `ShowInspector` function, and moved
it to a separate backCompat file. Now the static Inspector class is
where all back compat lives, and the new expanded Inspector v2 API is
all through the new `ShowInspector` function.
- Added support for various Inspector v1 options.

There is more left to do, but I want to do it in chunks so the PRs don't
get too big.

Also as the comment says, support for `debugLayer` is commented out for
now, as it would otherwise break the ability to dynamically switch
between Inspector v1 and v2 in Playground and Sandbox. It will be
enabled when we are ready to make v2 the default and are ok with a page
reload to switch between v1 and v2.