Hook up Tools.PerformanceLogLevel / Tools.StartPerformanceCounter / Tools.EndPerformanceCounter to Babylon Native counterparts (#17450)
NativeTracing has been partially implemented in Babylon Native for a
long time. These changes hook up the Babylon.js perf tracing related
APIs in Tools to the Babylon Native tracing APIs. This extra
functionality will only execute in the context of Babylon Native, and is
resilient to older versions of Babylon Native. For these changes to
fully work, they need
https://github.com/BabylonJS/BabylonNative/pull/1569.

Couple additional changes related to our native interfaces:
- I did a little bit of cleanup on the `INative` interface to make it
more type safe
- ~I added a declation for the `_native` global in
`INativeInterfaces.ts`, which makes it visible globally everywhere so it
doesn't have to be redeclared in every file it is used in. However, I
declared it as potentially undefined (since it is), so I didn't touch a
few files that already declared it as non nullable (like NativeEngine)
as there would be a ton of changes and I didn't want this PR to get too
overloaded.~ Never mind, not in the mood to battle the custom umd type
declaration generation script today.