Inspector v2: Back compat for explorerExtensibility v1 option (#17470)
Prior to this PR Inspector v2 supports Scene Explorer "commands" which
are shown inline in the tree view item, and as such they _must_ have an
icon. For the Inspector v1 `explorerExtensibility` option, it is
effectively a Scene Explorer command, but has no icon. In Inspector v1,
there is a `...` icon at the end that has the context menu. In Inspector
v2, we already have a right click context menu for other things like
expand/collapse. Given this, the strategy for Inspector v2 is:

- Allow for "inline" vs. "contextMenu" commands for Scene Explorer.
"inline" is the default and requires an icon, where as for "contextMenu"
the icon is optional.
- For commands that are toggle "type" and contextMenu "mode", if an icon
is provided it is used, otherwise a checkmark is used, but we don't
double up the iconography and have both a checkmark and an icon.

With this in place, `explorerExtensibility` is simply adding commands of
type "action" and mode "contextMenu" with no icon.

<img width="389" height="331" alt="image"
src="https://github.com/user-attachments/assets/4d54c40b-8011-4c9e-be98-2de5b7e96106"
/>