Inspector v2: Back compat for gizmoCamera option and Inspector.Popup* functions (#17483)
This PR implements back compat for the following:
- The `gizmoCamera` Inspector v1 option. There is already an
`IGizmoService` that exposes the utility layer, so it's just a matter of
calling `setRenderCamera` on that utility layer.
- The `PopupSceneExplorer`, `PopupInspector`, and `PopupEmbed` functions
on the Inspector v1 `Inspector` class. Since Inspector v2 allows side
panes to be moved around, this just maps to undocking either the left or
right side pane container, based on the *default* dock state of where
Scene Explorer is (left), where the other default Inspector v2 panes are
(right), and where all panes are in embedMode (right). To enable this, I
exposed some new API surface area on `IShellService` that allows
programmatically requesting the side pane containers to be
docked/undocked, and use that to implement the legacy/back compat Popup*
functions.