Inspector v2: Back compat for Inspector.OnSelectionChangeObservable and Inspector.OnPropertyChangedObservable (#17488)
This PR adds back compat imple for the following:
- `Inspector.OnSelectionChangeObservable` - this one is simply observing
the existing `ISelectionService.onSelectedEntityChanged`.
- `Inspector.OnPropertyChangedObservable` - for this one, I introduced a
React `PropertyContext` that is used in the Properties service. It
exposes an Observable for when properties change. The Properties service
exposes the observable, and `BoundProperty` consumes the context and
fires the observable when a property is changed. Any properties handled
outside of `BoundProperty` will not automatically fire the event, so we
need to finish moving any remaining direct usages of *PropertyLine
components to `BoundProperty`, and in any cases where we can't, we need
to manually fire the observable from the context. I will work on this
when I get back to finishing fleshing out the properties pane, but most
properties should already be covered as we are using `BoundProperty`
extensively already.