Smart Filters: Add const property support - consts in GLSL defined blocks that are settable at design time (#17479)
With this change, consts in fragment shaders of blocks defined using
GLSL can be annotated to be treated as properties on the resulting
block. Those properties can be set before runtime creation and then are
baked into the shader as consts. The properties are annotated for the
Smart Filter Editor to see and display in the property pane on the right
when a block instance is selected. The optimizer is updated to know that
such consts cannot be consolidated into a single const for all instances
of that block type, although regular unannotated consts will still be
consolidated.

The syntax for the annotation can either be:
```
// { "property": true }
const float mode = 1.0;
```
and this will show a text box in the Smart Filter Editor:
<img width="569" height="150" alt="image"
src="https://github.com/user-attachments/assets/c7b9f77f-2f0e-4018-9daf-c6067ba52139"
/>


Or you can supply a list of options for the value:
```
// { "property": { "options": { "Normal": 0, "Crazy": 1, "Odd": 2 } } }
const float mode = 1.0;
```
and this will show a drop down in the Smart Filter Editor:
<img width="550" height="155" alt="image"
src="https://github.com/user-attachments/assets/9f564e6d-57fe-422f-98f6-a3be16f7e619"
/>

This change also updates the serialization of CustomShaderBlocks to
include the values of these properties in the data section:
```
{
      "name": "Example",
      "uniqueId": 34,
      "blockType": "ConstPropertyExample",
      "namespace": "Test",
      "comments": null,
      "data": {
        "customProperties": [
          {
            "name": "mode",
            "value": 1
          }
        ]
      },
      "outputTextureOptions": {
        "ratio": 1,
        "format": 5,
        "type": 0
      }
    }
```

Note: this change is backwards compatible.

---------

Co-authored-by: AmoebaChant <kevbrown@microsoft.com>