Update playground's monaco-editor package to fix sourcemap errors
With the new playground the monaco-editor package has errors due to missing sourcemaps. This change takes the upstream fix for this and updates our code to fix the deprecation issues introduced by the update.

Note that the webpack `monacoEditorPath` option has to be set now to avoid new webpack errors occurring when building/running the playground.