Inspector v2: Back compat for Inspector.MarkLineContainerTitleForHighlighting (#17495)
This PR adds support for
`Inspector.MarkLineContainerTitleForHighlighting` and
`Inspector.MarkMultipleLineContainerTitlesForHighlighting`.
- Add a prop to `Accordion` to allow highlighting sections. Similar to
v1, the other sections are collapsed.
- Add a prop for an imperative ref for the `ExtensibleAccordion` that
provides an interface for triggering the highlighting in the contained
`Accordion`. Similar to v1, the highlighting state is cleared when a
different entity is selected.
- Add a `highlightSections` function to the `IPropertiesService` that
calls into the highlighting interface on the `ExtensibleAccordion`.
- Implement `Inspector.MarkMultipleLineContainerTitlesForHighlighting`
by providing a service that consumes the `IPropertiesService` and calls
`highlightSections`. There is also a giant mapping of Inspector v1
property section names to Inspector v2 property section names (thank you
GitHub Copilot Chat 😁).


https://github.com/user-attachments/assets/0aad1986-3b99-4767-952b-335b7a58e3e5