Node material: Fix properties not being parsed correctly (#17502)
I think this is a better fix than #17498 => even with the PR, there are
still some properties that are not parsed correctly (**comment**,
**ignoreAlpha** and **maxSimultaneousLights**).

In #17098, my mistake was to add
`SerializationHelper.ParseProperties(source, this, this.getScene(),
rootUrl);` in `loadFromSerialization` and not in
`parseSerializedObject`: when we call `parseSerializedObject` directly,
the fix was not in effect.

What do you think?