[GeospatialCamera] Update flyToAsync hop behavior to animate center with parabolic hop (vs scale radius) (#17512)
Rather than using customKeys for customizing the interpolating behavior,
allow user to customize any part of the animation via callback.

This is now used to animate center rather than radius so that center can
have a parabolic hop animation based on passed in scale factor.

The problem with previous approach was that just animating radius is not
super noticeable if the camera is already very close to ground /
pitchedtowards horizon. This approach creates a true parabolic arc
motion

#17451

Co-authored-by: Georgina Halpern <gehalper@microsoft.com>