Inspector v2: Back compat for "attaching" UMD globals and Scene.debugLayer (#17521)
This change is mostly about making it possible to "attach/detach"
Inspector v2 to the global scope (e.g. `globalThis.BABYLON.Inspector`
for the class and `globalThis.INSPECTOR` for the module) and
`Scene.debugLayer` *dynamically* via explicit function calls, while also
calling those `Attach*` functions by default as side effects. This
enables the following:
- Side effect back compat.
- Ability to "detach" dynamically, which will enable swapping between
Inspector v1 and v2 at runtime, which we still need to be able to do
until the 9.0 release. This will let us switch to using the UMD package
to support versioning (e.g. using an older version of the platform,
including an older version of Inspector v2) which by default will
override the globals and Scene.debugLayer, but we can dynamically detach
to restore v1.
- A future way to be able to explicitly "attach" without relying on side
effects, which we'll want when we get to all the side effects work
anyway.

I also snuck in a couple of unrelated small changes:
- Remove the use of makeStyles in the inline "Canvas Injector" service
definition. The style definition must be determined inside the function
because it is dynamic and depends on the `canvasContainerDisplay`, but
`makeStyles` incorrectly detects that it is being used within a function
component and emits a warning. Since the style is tiny and won't be
re-used, just making it an inline React style is fine and gets rid of
the annoying runtime warning.
- In the Inspector v2 test app, add a `div` layer so that React doesn't
mount directly to the body. This gets rid of a distracting "error"
displayed in the console for the test app.